REMOVED: Non-necessary casts FIXED: Crash in Spell::Heal where unitTarget would not be of Player type; anyway changing unitTarget to u_caster as it gives more sense (KFL please check if its still what you wanted)

+++ src/arcemu-world/SpellEffects.cpp
@@ -2445,7 +2445,7 @@ void Spell::SpellEffectHeal(uint32 i) // Heal
 				}
 			}break;
 		default:
-			Heal((int32)damage);
+			Heal(damage);
 			break;
 		}
 	}
